some proposals
