+++ package.json
@@ -10,7 +10,8 @@
     "types": "dist/lib/index.d.ts",
     "scripts": {
         "test": "node dist/test/",
-        "build": "tsc",
+        "build": "tsc && yarn copyfiles",
+        "copyfiles": "copyfiles -u 1 src/**/*.svg dist/",
         "lint:check": "eslint . --ext .ts,.tsx",
         "lint": "npm run lint:check -- --fix",
         "_format": "prettier '**/*.{ts,tsx,json,md}'",
@@ -62,6 +63,7 @@
         "@types/react": "^17.0.4",
         "@typescript-eslint/eslint-plugin": "^4.15.1",
         "@typescript-eslint/parser": "^4.15.1",
+        "copyfiles": "^2.4.1",
         "eslint": "^7.20.0",
         "eslint-config-prettier": "^7.2.0",
         "evt": "^2.0.0-beta.18",
