+++ src/lib/components/HomepageTemplate/HomepageTemplate.tsx
@@ -1,3 +1,4 @@
+import { useEffect } from "react";
 import { Banner } from "./Banner";
 import { MainSection } from "./MainSection";
 import { ReviewSlider } from "./ReviewSlider";
@@ -7,7 +8,7 @@ import type { Props as MainSectionProps } from "./MainSection";
 import type { Props as FooterProps } from "./Footer";
 import type { Props as ReviewSliderProps } from "./ReviewSlider";
 
-export type Props = {
+export type HomepageTemplate = {
     banner?: BannerProps;
     /**
      * enter the assets in an array.
@@ -52,8 +53,21 @@ export type Props = {
     footer?: FooterProps;
 };
 
-export const App = (props: Props) => {
+export const HomepageTemplate = (props: HomepageTemplate) => {
     const { footer, banner, mainSection, reviewSlider } = props;
+
+    useEffect(() => {
+        const script = document.createElement("script");
+
+        script.src = "https://buttons.github.io/buttons.js";
+        script.async = true;
+
+        document.body.appendChild(script);
+
+        return () => {
+            document.body.removeChild(script);
+        };
+    }, []);
 
     return (
         <>
