+++ src/lib/global.d.ts
@@ -1,0 +1,4 @@
+declare module "*.svg" {
+    const _default: string;
+    export default _default;
+}
