+++ src/lib/index.ts
@@ -1,0 +1,5 @@
+export * from "./components/HomepageTemplate";
+
+import onyxiaLogoUrl from "./assets/svg/OnyxiaLogo.svg";
+
+export { onyxiaLogoUrl };
